requirements.txt
# Data Processing
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
pyyaml>=6.0
tqdm>=4.65.0
jieba>=0.42.1
pandas>=2.0.0

# LLM Training
transformers>=4.40.0
datasets>=2.14.0
accelerate>=0.25.0
peft>=0.10.0
bitsandbytes>=0.41.0
llamafactory>=0.9.0

# Inference
sentencepiece>=0.1.99
protobuf>=3.20.0
